chore(base_layer): add test-utils + move ganache
This will be used in the L1 provider.

MOVE only, no logic changes.
